requirements.txt
fastapi
uvicorn[standard]
boto3
requests
python-dotenv